move supporting utils for TM paper to htmresearch/support
