+++ projects/sequence_prediction/continuous_sequence/plot.py
@@ -26,123 +26,15 @@ from matplotlib import pyplot as plt
 plt.ion()
 
 from suite import Suite
-from errorMetrics import *
+from htmresearch.support.sequence_learning_utils import *
 import pandas as pd
 import numpy as np
-
-def movingAverage(a, n):
-  movingAverage = []
-
-  for i in xrange(len(a)):
-    start = max(0, i - n)
-    end = i+1
-    #
-    # start = i
-    # end = min(len(a), i + n)
-
-    # start = max(0, i-n/2)
-    # end = min(len(a), i+n/2)
-    #
-    values = a[start:end]
-    movingAverage.append(sum(values) / float(len(values)))
-
-  return movingAverage
-
 
 
 def plotMovingAverage(data, window, label=None):
   movingData = movingAverage(data, min(len(data), window))
   style = 'ro' if len(data) < window else ''
   plt.plot(range(len(movingData)), movingData, style, label=label)
-
-
-
-def plotAccuracy(results, truth, train=None, window=100, label=None, params=None, errorType=None):
-  plt.title('Prediction Error Over Time')
-
-  error = results[0]
-
-  x = results[1]
-  x = x[:len(error)]
-
-  # print results
-  # print params['compute_after']
-  # if params is not None:
-  #   error[np.where(x < params['compute_after'])[0]] = np.nan
-
-  error[:5904] = np.nan
-
-  movingData = movingAverage(error, min(len(error), window))
-
-  if errorType == 'square_deviation':
-    print label, " Avg NRMSE:", np.sqrt(np.nanmean(error))/np.nanstd(truth)
-    meanError = np.sqrt(np.nanmean(error))/np.nanstd(truth)
-    avgError = np.sqrt(np.array(movingData))/np.nanstd(truth)
-  elif errorType == 'negLL':
-    print label, " Avg negLL:", np.nanmean(error)
-    meanError = np.nanmean(error)
-    avgError = movingData
-  elif errorType == 'mape':
-
-    normFactor = np.nanstd(truth)
-    print label, " MAPE:", np.nanmean(error)  / normFactor
-    meanError = np.nanmean(error) / normFactor
-    avgError = movingData / normFactor
-  else:
-    raise NotImplementedError
-
-  plt.plot(x, avgError, label=label)
-  plt.xlabel("# of elements seen")
-  plt.ylabel("{0} over last {1} record".format(errorType, window))
-  if train is not None:
-    for i in xrange(len(train)):
-      if train[i]:
-        plt.axvline(x[i], color='orange')
-
-  if params is not None:
-    if params['perturb_after'] < len(x):
-      plt.axvline(x[params['perturb_after']], color='black', linestyle='--')
-
-  plt.xlim(x[0], x[len(x)-1])
-  return error
-
-  # plt.ylim(0, 1.001)
-
-def computeAltMAPE(truth, prediction, startFrom=0):
-  return np.nanmean(np.abs(truth[startFrom:] - prediction[startFrom:]))/np.nanmean(np.abs(truth[startFrom:]))
-
-
-def computeNRMSE(truth, prediction, startFrom=0):
-  squareDeviation = computeSquareDeviation(prediction, truth)
-  squareDeviation[:startFrom] = None
-  return np.sqrt(np.nanmean(squareDeviation))/np.nanstd(truth)
-
-
-def computeSquareDeviation(predictions, truth):
-
-  squareDeviation = np.square(predictions-truth)
-  return squareDeviation
-
-
-def computeLikelihood(predictions, truth, encoder):
-  targetDistribution = np.zeros(predictions.shape)
-  for i in xrange(len(truth)):
-    if not np.isnan(truth[i]) and truth is not None:
-      targetDistribution[i, :] = encoder.encode(truth[i])
-
-  # calculate negative log-likelihood
-  Likelihood = np.multiply(predictions, targetDistribution)
-  Likelihood = np.sum(Likelihood, axis=1)
-
-  minProb = 0.01
-  Likelihood[np.where(Likelihood < minProb)[0]] = minProb
-  negLL = -np.log(Likelihood)
-
-  return negLL
-
-
-def computeAbsouteError(predictions, truth):
-  return np.abs( (predictions-truth))
