+++ projects/sequence_prediction/continuous_sequence/run_elm.py
@@ -31,7 +31,7 @@ from swarm_runner import SwarmRunner
 from scipy import random
 
 import pandas as pd
-from errorMetrics import *
+from htmresearch.support.sequence_learning_utils import *
 
 from htmresearch.algorithms.online_extreme_learning_machine import OSELM
