+++ projects/sequence_prediction/continuous_sequence/run_lstm_scalarEncoder.py
@@ -36,7 +36,7 @@ from swarm_runner import SwarmRunner
 
 
 import pandas as pd
-from errorMetrics import *
+from htmresearch.support.sequence_learning_utils import *
 
 from nupic.encoders.scalar import ScalarEncoder
