+++ projects/sequence_prediction/continuous_sequence/run_tdnn.py
@@ -30,7 +30,7 @@ from swarm_runner import SwarmRunner
 from scipy import random
 
 import pandas as pd
-from errorMetrics import *
+from htmresearch.support.sequence_learning_utils import *
 
 from pybrain.tools.shortcuts import buildNetwork
 from pybrain.supervised import BackpropTrainer
